requirements.txt
black==23.11.0
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
Flask==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==4.1.0
pycparser==2.21
pyOpenSSL==23.3.0
signxml==3.2.1
Werkzeug==3.0.1